work around limitation when building via ssh
build via ssh on some platform hang in some test and/or misbehave in
general.
This work around it by allowing to setup a little loopback server
that need to be started in a regular desktop session.
then connect via ssh and use the client to trigger a make
the server will pick up the request and run the make in the proper
environment.

The server is meant to run with LODE_HOME defined and will only
handle simple command without argument (make and test as of now)
and will also verify that the working directory to run the command is
a sub-directory of LODE_HOME