Fix installation of python2 and python3 on windows
Due to the dependency algorithm in setup, python2 was only installed
as a dependency to git. Python3 was not installed

This was tested on a 64 bit system, by deleting cygwin and rerunning
the script.

Fix by putting python,python3 in a seperate -P line after the
others, this caused setup to detect the need to install both
python(2) and python3.

Normally the sequence should not matter, however with cygwin
(and some linux distros) this is not the case.

Change-Id: I76562009d896d42fe683eda4c2863c247cadd4ed
Reviewed-on: https://gerrit.libreoffice.org/19696
Reviewed-by: David Ostrovsky <david@ostrovsky.org>
Reviewed-by: Norbert Thiebaud <nthiebaud@gmail.com>
Tested-by: Norbert Thiebaud <nthiebaud@gmail.com>